Switch to iOS WKWebView
This is now required for submission to the app store.
History:
- I first followed the basic cordova instructions
    https://cordova.apache.org/howto/2020/03/18/wkwebviewonly.html
- Ran into issues with file:// XHR loading. Attempted to fix that using
    https://github.com/oracle/cordova-plugin-wkwebview-file-xhr
- This worked great for everything other than the translations. Apparently, the
  app was not able to intercept the translation file loads - when I added logs
  for which files were intercepted, the translation files were not on the list.
- So I then added logs to the translation file loading code and went back to UIWebView
  to see when they were called, so I could walk back up the chain to figure out
  what was going on.
- The translation file loading code did not print anything out on UIWebView as well!
- I then tried the ionic wkwebview which, after clearing out all packages and
  platforms and rebuilding, worked out of the box.

Yay!?

Other changes are mainly around trying to get the new UI to not bleed into the
edges of the screen. It is not perfect, but it is much better, at least at the
top. I couldn't get the bottom to work after ~ 1 hour of testing, so leaving it
for somebody else to clean up.